move solana logic
