+++ deployment/ccip/changeset/solana/cs_solana_token_test.go
@@ -1,4 +1,4 @@
-package changeset_test
+package changeset_solana_test
 
 import (
 	"context"
@@ -12,7 +12,9 @@ import (
 
 	solTokenUtil "github.com/smartcontractkit/chainlink-ccip/chains/solana/utils/tokens"
 	"github.com/smartcontractkit/chainlink/deployment"
-	"github.com/smartcontractkit/chainlink/deployment/common/changeset"
+	changeset_solana "github.com/smartcontractkit/chainlink/deployment/ccip/changeset/solana"
+	"github.com/smartcontractkit/chainlink/deployment/ccip/changeset/testhelpers"
+	commonchangeset "github.com/smartcontractkit/chainlink/deployment/common/changeset"
 	"github.com/smartcontractkit/chainlink/deployment/environment/memory"
 	"github.com/smartcontractkit/chainlink/v2/core/logger"
 )
@@ -24,10 +26,10 @@ func TestDeploySolanaToken(t *testing.T) {
 		SolChains: 1,
 	})
 	solChain1 := e.AllChainSelectorsSolana()[0]
-	e, err := changeset.ApplyChangesets(t, e, nil, []changeset.ChangesetApplication{
+	e, err := commonchangeset.ApplyChangesets(t, e, nil, []commonchangeset.ChangesetApplication{
 		{
-			Changeset: changeset.WrapChangeSet(changeset.DeploySolanaToken),
-			Config: &changeset.DeploySolanaTokenConfig{
+			Changeset: commonchangeset.WrapChangeSet(changeset_solana.DeploySolanaToken),
+			Config: &changeset_solana.DeploySolanaTokenConfig{
 				ChainSelector:    solChain1,
 				TokenName:        "spl-token-2022",
 				TokenProgramName: "spl-token-2022",
@@ -37,8 +39,8 @@ func TestDeploySolanaToken(t *testing.T) {
 			},
 		},
 		{
-			Changeset: changeset.WrapChangeSet(changeset.MintSolanaToken),
-			Config: &changeset.MintSolanaTokenConfig{
+			Changeset: commonchangeset.WrapChangeSet(changeset_solana.MintSolanaToken),
+			Config: &changeset_solana.MintSolanaTokenConfig{
 				ChainSelector: solChain1,
 				TokenName:     "spl-token-2022",
 				TokenProgram:  "spl-token-2022",
@@ -60,4 +62,8 @@ func TestDeploySolanaToken(t *testing.T) {
 	require.NoError(t, err)
 	require.Equal(t, int(1000), outVal)
 	require.Equal(t, 9, int(outDec))
+}
+
+func TestDeployLinkToken(t *testing.T) {
+	testhelpers.DoDeployLinkToken(t, 1)
 }
